Firefox compatibility

+++ src/gladys-gateway-js/lib/helpers.js
@@ -1,10 +1,10 @@
 function ab2str(buf) {
-  return String.fromCharCode.apply(null, new Uint8Array(buf));
+  return String.fromCharCode.apply(null, new Uint16Array(buf));
 }
 
 function str2ab(str) {
   let buf = new ArrayBuffer(str.length * 2);
-  let bufView = new Uint8Array(buf);
+  let bufView = new Uint16Array(buf);
   for (let i = 0, strLen = str.length; i < strLen; i++) {
     bufView[i] = str.charCodeAt(i);
   }
